# bdd/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Mwague backend", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/office/transactions": {
      "get": {
        "tags": ["transactions"],
        "summary": "Get Office Transactions",
        "description": "get all transactions for an office",
        "operationId": "get_office_transactions_api_v1_office_transactions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "type": "array",
                  "title": "Response Get Office Transactions Api V1 Office Transactions Get"
                }
              }
            }
          }
        },
        "security": [{ "OAuth2AuthorizationCodeBearer": [] }]
      }
    },
    "/api/v1/agent/{initials}/transactions": {
      "get": {
        "tags": ["transactions"],
        "summary": "Get Agent Transactions",
        "description": "get all transactions for an agent",
        "operationId": "get_agent_transactions_api_v1_agent__initials__transactions_get",
        "parameters": [
          {
            "required": true,
            "schema": { "type": "string", "title": "Initials" },
            "name": "initials",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "type": "array",
                  "title": "Response Get Agent Transactions Api V1 Agent  Initials  Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2AuthorizationCodeBearer": [] }]
      }
    },
    "/api/v1/transaction": {
      "post": {
        "tags": ["transactions"],
        "summary": "Request Transaction",
        "description": "request a transaction for approval, this will just created the transaction in the db",
        "operationId": "request_transaction_api_v1_transaction_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransactionRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransactionResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2AuthorizationCodeBearer": [] }]
      }
    },
    "/api/v1/transaction/{transaction_code}/review": {
      "post": {
        "tags": ["transactions"],
        "summary": "Review Transaction",
        "description": "review a transaction request",
        "operationId": "review_transaction_api_v1_transaction__transaction_code__review_post",
        "parameters": [
          {
            "required": true,
            "schema": { "type": "string", "title": "Transaction Code" },
            "name": "transaction_code",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransactionReviewReq" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransactionResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2AuthorizationCodeBearer": ["office_admin"] }]
      }
    },
    "/api/v1/transaction/{code}": {
      "get": {
        "tags": ["transactions"],
        "summary": "Get Office Transactions With Details",
        "description": "get all transactions for an office with details",
        "operationId": "get_office_transactions_with_details_api_v1_transaction__code__get",
        "parameters": [
          {
            "required": true,
            "schema": { "type": "string", "title": "Code" },
            "name": "code",
            "in": "path"
          },
          {
            "required": true,
            "schema": { "$ref": "#/components/schemas/TransactionType" },
            "name": "tr_type",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/Internal" },
                    { "$ref": "#/components/schemas/Deposit" },
                    { "$ref": "#/components/schemas/Sending" },
                    { "$ref": "#/components/schemas/ForEx" },
                    { "$ref": "#/components/schemas/External" }
                  ],
                  "title": "Response Get Office Transactions With Details Api V1 Transaction  Code  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2AuthorizationCodeBearer": ["office_admin"] }]
      },
      "put": {
        "tags": ["transactions"],
        "summary": "Update Transaction",
        "description": "update a transaction",
        "operationId": "update_transaction_api_v1_transaction__code__put",
        "parameters": [
          {
            "required": true,
            "schema": { "type": "string", "title": "Code" },
            "name": "code",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransactionRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransactionResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2AuthorizationCodeBearer": [] }]
      }
    },
    "/api/v1/transaction/{code}/pay": {
      "post": {
        "tags": ["transactions"],
        "summary": "Add Payment",
        "description": "add payment to a transaction",
        "operationId": "add_payment_api_v1_transaction__code__pay_post",
        "parameters": [
          {
            "required": true,
            "schema": { "type": "string", "title": "Code" },
            "name": "code",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PaymentRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2AuthorizationCodeBearer": [] }]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Amount": {
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount",
            "strict": true
          },
          "rate": {
            "type": "number",
            "minimum": 0.0,
            "title": "Rate",
            "strict": true
          }
        },
        "type": "object",
        "required": ["amount", "rate"],
        "title": "Amount",
        "description": "Amount and rate of a transaction."
      },
...
      "Currency": {
        "enum": ["USD", "EUR", "AED", "CFA", "GNF", "RMB"],
        "title": "Currency",
        "description": "An enumeration."
      },
...
      "Deposit": {
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount",
            "strict": true
          },
          "rate": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Rate",
            "strict": true
          },
          "code": { "type": "string", "maxLength": 64, "title": "Code" },
          "state": { "$ref": "#/components/schemas/TransactionState" },
          "type": { "$ref": "#/components/schemas/TransactionType" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "office_id": {
            "type": "string",
            "format": "uuid",
            "title": "Office Id"
          },
          "org_id": { "type": "string", "format": "uuid", "title": "Org Id" },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "reviwed_by": {
            "type": "string",
            "format": "uuid",
            "title": "Reviwed By"
          },
          "history": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "History",
            "default": {}
          },
          "notes": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "Notes",
            "default": {}
          },
          "owner_initials": { "type": "string", "title": "Owner Initials" }
        },
        "type": "object",
        "required": [
          "amount",
          "rate",
          "code",
          "state",
          "type",
          "office_id",
          "org_id",
          "created_by",
          "owner_initials"
        ],
        "title": "Deposit",
        "description": "Transaction database model"
      },
      "DepositRequest": {
        "properties": {
          "type": { "type": "string", "enum": ["DEPOSIT"], "title": "Type" },
          "receiver": { "type": "string", "title": "Receiver" }
        },
        "type": "object",
        "required": ["type", "receiver"],
        "title": "DepositRequest"
      },
...
      "External": {
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount",
            "strict": true
          },
          "rate": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Rate",
            "strict": true
          },
          "code": { "type": "string", "maxLength": 64, "title": "Code" },
          "state": { "$ref": "#/components/schemas/TransactionState" },
          "type": { "$ref": "#/components/schemas/TransactionType" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "office_id": {
            "type": "string",
            "format": "uuid",
            "title": "Office Id"
          },
          "org_id": { "type": "string", "format": "uuid", "title": "Org Id" },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "reviwed_by": {
            "type": "string",
            "format": "uuid",
            "title": "Reviwed By"
          },
          "history": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "History",
            "default": {}
          },
          "notes": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "Notes",
            "default": {}
          },
          "sender_initials": { "type": "string", "title": "Sender Initials" },
          "charges": { "type": "number", "minimum": 0.0, "title": "Charges" },
          "customer": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "Customer",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "amount",
          "rate",
          "code",
          "state",
          "type",
          "office_id",
          "org_id",
          "created_by",
          "sender_initials",
          "charges"
        ],
        "title": "External",
        "description": "Transaction database model"
      },
      "ExternalRequest": {
        "properties": {
          "type": { "type": "string", "enum": ["EXTERNAL"], "title": "Type" },
          "sender": { "type": "string", "title": "Sender" },
          "customer": { "$ref": "#/components/schemas/CustomerDetails" },
          "payment_currency": { "$ref": "#/components/schemas/Currency" }
        },
        "type": "object",
        "required": ["type", "sender", "payment_currency"],
        "title": "ExternalRequest"
      },
      "ForEx": {
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount",
            "strict": true
          },
          "rate": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Rate",
            "strict": true
          },
          "code": { "type": "string", "maxLength": 64, "title": "Code" },
          "state": { "$ref": "#/components/schemas/TransactionState" },
          "type": { "$ref": "#/components/schemas/TransactionType" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "office_id": {
            "type": "string",
            "format": "uuid",
            "title": "Office Id"
          },
          "org_id": { "type": "string", "format": "uuid", "title": "Org Id" },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "reviwed_by": {
            "type": "string",
            "format": "uuid",
            "title": "Reviwed By"
          },
          "history": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "History",
            "default": {}
          },
          "notes": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "Notes",
            "default": {}
          },
          "bid_rate": { "type": "number", "title": "Bid Rate" },
          "offer_rate": { "type": "number", "title": "Offer Rate" },
          "method": { "$ref": "#/components/schemas/PaymentMethod" }
        },
        "type": "object",
        "required": [
          "amount",
          "rate",
          "code",
          "state",
          "type",
          "office_id",
          "org_id",
          "created_by",
          "bid_rate",
          "offer_rate",
          "method"
        ],
        "title": "ForEx",
        "description": "Une transaction de change est effectu\u00e9"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Internal": {
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount",
            "strict": true
          },
          "rate": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Rate",
            "strict": true
          },
          "code": { "type": "string", "maxLength": 64, "title": "Code" },
          "state": { "$ref": "#/components/schemas/TransactionState" },
          "type": { "$ref": "#/components/schemas/TransactionType" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "office_id": {
            "type": "string",
            "format": "uuid",
            "title": "Office Id"
          },
          "org_id": { "type": "string", "format": "uuid", "title": "Org Id" },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "reviwed_by": {
            "type": "string",
            "format": "uuid",
            "title": "Reviwed By"
          },
          "history": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "History",
            "default": {}
          },
          "notes": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "Notes",
            "default": {}
          },
          "sender_initials": { "type": "string", "title": "Sender Initials" },
          "receiver_initials": {
            "type": "string",
            "title": "Receiver Initials"
          },
          "charges": { "type": "number", "minimum": 0.0, "title": "Charges" }
        },
        "type": "object",
        "required": [
          "amount",
          "rate",
          "code",
          "state",
          "type",
          "office_id",
          "org_id",
          "created_by",
          "sender_initials",
          "receiver_initials",
          "charges"
        ],
        "title": "Internal",
        "description": "Transaction database model"
      },
      "InternalRequest": {
        "properties": {
          "type": { "type": "string", "enum": ["INTERNAL"], "title": "Type" },
          "sender": { "type": "string", "title": "Sender" },
          "receiver": { "type": "string", "title": "Receiver" }
        },
        "type": "object",
        "required": ["type", "sender", "receiver"],
        "title": "InternalRequest",
        "description": "Internal transaction request."
      },
...
      "NoteList": {
        "properties": {
          "notes": {
            "items": { "$ref": "#/components/schemas/Note" },
            "type": "array",
            "title": "Notes",
            "default": []
          }
        },
        "type": "object",
        "title": "NoteList"
      },
...
      "PaymentMethod": {
        "enum": ["CASH", "BANK", "MOBILE"],
        "title": "PaymentMethod",
        "description": "An enumeration."
      },
      "PaymentRequest": {
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount",
            "strict": true
          },
          "payment_type": { "$ref": "#/components/schemas/TransactionType" },
          "notes": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "Notes",
            "default": {}
          }
        },
        "type": "object",
        "required": ["amount", "payment_type"],
        "title": "PaymentRequest"
      },
      "PaymentResponse": {
        "properties": {
          "payment_date": {
            "type": "string",
            "format": "date",
            "title": "Payment Date"
          },
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount",
            "strict": true
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "transaction_type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "state": { "$ref": "#/components/schemas/PaymentState" },
          "notes": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "Notes",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "payment_date",
          "amount",
          "transaction_id",
          "transaction_type",
          "state"
        ],
        "title": "PaymentResponse"
      },
      "PaymentState": {
        "enum": [1, 2],
        "title": "PaymentState",
        "description": "An enumeration."
      },
...
      "Sending": {
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount",
            "strict": true
          },
          "rate": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Rate",
            "strict": true
          },
          "code": { "type": "string", "maxLength": 64, "title": "Code" },
          "state": { "$ref": "#/components/schemas/TransactionState" },
          "type": { "$ref": "#/components/schemas/TransactionType" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "office_id": {
            "type": "string",
            "format": "uuid",
            "title": "Office Id"
          },
          "org_id": { "type": "string", "format": "uuid", "title": "Org Id" },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "reviwed_by": {
            "type": "string",
            "format": "uuid",
            "title": "Reviwed By"
          },
          "history": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "History",
            "default": {}
          },
          "notes": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "Notes",
            "default": {}
          },
          "receiver_initials": {
            "type": "string",
            "title": "Receiver Initials"
          },
          "bid_rate": { "type": "number", "title": "Bid Rate" },
          "offer_rate": { "type": "number", "title": "Offer Rate" },
          "method": { "$ref": "#/components/schemas/PaymentMethod" },
          "payment_currency": { "$ref": "#/components/schemas/Currency" },
          "charges": { "type": "number", "minimum": 0.0, "title": "Charges" },
          "customer_sender": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "Customer Sender",
            "default": {}
          },
          "customer_receiver": {
            "additionalProperties": { "anyOf": [{ "type": "object" }, {}] },
            "type": "object",
            "title": "Customer Receiver",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "amount",
          "rate",
          "code",
          "state",
          "type",
          "office_id",
          "org_id",
          "created_by",
          "receiver_initials",
          "bid_rate",
          "offer_rate",
          "method",
          "payment_currency",
          "charges"
        ],
        "title": "Sending",
        "description": "Transaction database model"
      },
      "SendingRequest": {
        "properties": {
          "type": { "type": "string", "enum": ["SENDING"], "title": "Type" },
          "receiver_initials": {
            "type": "string",
            "title": "Receiver Initials"
          },
          "customer_sender": { "$ref": "#/components/schemas/CustomerDetails" },
          "customer_receiver": {
            "$ref": "#/components/schemas/CustomerDetails"
          },
          "bid_rate": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Bid Rate",
            "strict": true
          },
          "offer_rate": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Offer Rate",
            "strict": true
          },
          "payment_method": { "$ref": "#/components/schemas/PaymentMethod" },
          "payment_currency": { "$ref": "#/components/schemas/Currency" }
        },
        "type": "object",
        "required": [
          "type",
          "receiver_initials",
          "bid_rate",
          "offer_rate",
          "payment_method",
          "payment_currency"
        ],
        "title": "SendingRequest"
      },
      "TransactionRequest": {
        "properties": {
          "currency": { "$ref": "#/components/schemas/Currency" },
          "amount": { "$ref": "#/components/schemas/Amount" },
          "charges": { "$ref": "#/components/schemas/Amount" },
          "transaction_type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "data": {
            "oneOf": [
              { "$ref": "#/components/schemas/InternalRequest" },
              { "$ref": "#/components/schemas/DepositRequest" },
              { "$ref": "#/components/schemas/ExternalRequest" },
              { "$ref": "#/components/schemas/SendingRequest" }
            ],
            "title": "Data",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "INTERNAL": "#/components/schemas/InternalRequest",
                "DEPOSIT": "#/components/schemas/DepositRequest",
                "EXTERNAL": "#/components/schemas/ExternalRequest",
                "SENDING": "#/components/schemas/SendingRequest"
              }
            }
          }
        },
        "type": "object",
        "required": ["currency", "amount"],
        "title": "TransactionRequest"
      },
      "TransactionResponse": {
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount",
            "strict": true
          },
          "rate": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Rate",
            "strict": true
          },
          "code": { "type": "string", "maxLength": 64, "title": "Code" },
          "state": { "$ref": "#/components/schemas/TransactionState" },
          "type": { "$ref": "#/components/schemas/TransactionType" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "charges": { "type": "number", "title": "Charges" },
          "notes": { "$ref": "#/components/schemas/NoteList" }
        },
        "type": "object",
        "required": ["amount", "rate", "code", "state", "type", "notes"],
        "title": "TransactionResponse"
      },
      "TransactionReviewReq": {
        "properties": {
          "currency": { "$ref": "#/components/schemas/Currency" },
          "amount": { "$ref": "#/components/schemas/Amount" },
          "charges": { "$ref": "#/components/schemas/Amount" },
          "transaction_type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "data": {
            "oneOf": [
              { "$ref": "#/components/schemas/InternalRequest" },
              { "$ref": "#/components/schemas/DepositRequest" },
              { "$ref": "#/components/schemas/ExternalRequest" },
              { "$ref": "#/components/schemas/SendingRequest" }
            ],
            "title": "Data",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "INTERNAL": "#/components/schemas/InternalRequest",
                "DEPOSIT": "#/components/schemas/DepositRequest",
                "EXTERNAL": "#/components/schemas/ExternalRequest",
                "SENDING": "#/components/schemas/SendingRequest"
              }
            }
          },
          "code": { "type": "string", "title": "Code" },
          "type": { "$ref": "#/components/schemas/TransactionType" },
          "state": {
            "allOf": [{ "$ref": "#/components/schemas/ValidationState" }],
            "nullable": false
          },
          "notes": { "type": "string", "title": "Notes" }
        },
        "type": "object",
        "required": ["currency", "amount", "code", "type", "state"],
        "title": "TransactionReviewReq"
      },
      "TransactionState": {
        "enum": ["REVIEW", "PENDING", "PAID", "CANCELLED", "REJECTED"],
        "title": "TransactionState",
        "description": "An enumeration."
      },
      "TransactionType": {
        "enum": ["DEPOSIT", "INTERNAL", "EXTERNAL", "SENDING", "FOREX"],
        "title": "TransactionType",
        "description": "An enumeration."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "ValidationState": {
        "enum": ["APPROVED", "REJECTED", "CANCELLED"],
        "title": "ValidationState",
        "description": "An enumeration."
      }
    },
    "securitySchemes": {
      "OAuth2AuthorizationCodeBearer": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "scopes": {},
            "authorizationUrl": "http://keycloak:8443/auth/realms/mwague/protocol/openid-connect/auth",
            "tokenUrl": "http://keycloak:8443/auth/realms/mwague/protocol/openid-connect/token"
          }
        }
      }
    }
  }
}